Add commandStart/End to ExecuteProgress
The main difficulty is the fact that the interface is defined in the tasks sub-project and has no access to sbt.State. To be able to make sense of what commands are doing, we need to thread the state before and after, and adding that as a second type parameter seems to work. However:

- this change leaks in many places
- it breaks source compatibility, so existing plugins and builds that used this interface will be broken

This is probably not the right way forward.